requirement.txt
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
customtkinter==0.3
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2
django-cors-headers==4.0.0
django-filter==24.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.0
fonttools==4.47.0
fqdn==1.5.1
idna==3.4
ipykernel==6.27.1
ipython==8.18.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nbclient==0.9.0
nbconvert==7.13.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.3
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
Pillow==9.5.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.7
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==1.7.1
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
razorpay==1.3.0
referencing==0.32.0
requests==2.31.0
rest-framework-simplejwt==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.15.2
seaborn==0.13.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.3
stack-data==0.6.3
terminado==0.18.0
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.0
types-python-dateutil==2.8.19.14
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.4
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
xlwt==1.3.0
